millau support pangolin message, migration code to primitives
